SWT-320 Rename **sweety** files back
