+++ src/Impulse.ts
