+++ tests/useWatchImpulse/illegal-usage.spec.tsx
